[minor] Change [faultHandler] stub so that implicits are inferable.
As Delphine had noted, she strangely needed to supply the usually
implicit [T] parameter in a few places.  This was because the code
stub for [faultHandler] was an [Admit].  Now it's actual code, which
fixes the inferences problem.

Still need to replace it with the real code, and prove correctness ...