[minor] Start defining pico machine executable semantics.
NB The code is really sad now, and missing calls and returns, but I'm
able to prove the correspondence theorem so far.

Coq buddies, I can't get (e)auto to use the step_rules constructors as
hints, even though they can be applied very straightforward in the end
of the proof. Any idea on how to fix that?