Remove requirement that tags are [handlerTag] in [tim_intro].
Fixed up Delphine's proofs, so hopefully this is not a problem.  If
it is a problem, then there is another route to go, which seems true
of the current definitions: the alternative is to make manifest in
the [handler_correct] premises that the initial memory has contents
with tags equal to [handlerTag].  This is true because [build_cache]
sets up the memory that way.