Prove [storeAt_spec_wp].
This was the last (used) conjecture, but there are still three
[admit]s in the proof of [handler_correct_succeed].  Two of them,
concerning valid addresses, are trivial, but will require some minor
interface changes (e.g. in [cache_hit], we could say that the memory
has at least 7 entries :P). The third concerns an "obvious" property:
that the fault handler code is in the fault handler memory when we
start, but APT's comment indicates this was not so obvious to prove.